AND-16244: Refactor switch case non-final resource id batch 3

+++ app/src/main/java/mega/privacy/android/app/main/megachat/chatAdapters/MegaChipChatExplorerAdapter.java
@@ -166,11 +166,8 @@ public class MegaChipChatExplorerAdapter extends RecyclerView.Adapter<MegaChipCh
                 Timber.w("Current position error - not valid value");
                 return;
             }
-            switch (view.getId()) {
-                case R.id.delete_icon_chip: {
-                    ((ChatExplorerFragment) fragment).deleteItemPosition(currentPosition);
-                    break;
-                }
+            if (view.getId() == R.id.delete_icon_chip) {
+                ((ChatExplorerFragment) fragment).deleteItemPosition(currentPosition);
             }
         } else {
             Timber.w("Error. Holder is Null");
